requirements.txt
asgiref==3.5.2
attrs==22.1.0
Django==4.1.3
django-filter==22.1
djangorestframework==3.14.0
exceptiongroup==1.0.4
factory-boy==3.2.1
Faker==15.3.3
flake8==6.0.0
iniconfig==1.1.1
isort==5.11.1
Markdown==3.4.1
mccabe==0.7.0
model-bakery==1.9.0
packaging==21.3
pluggy==1.0.0
psycopg2==2.9.5
pycodestyle==2.10.0
pyflakes==3.0.1
pyparsing==3.0.9
pytest==7.2.0
pytest-django==4.5.2
pytest-mock==3.10.0
python-dateutil==2.8.2
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
